Refactor to replace Float with Double for ContinuousConcept (#75)
* Refactor to replace Float with Double for ContinuousConcept
Replaced `Float` with `Double` in `ContinuousConcept` and related methods for better precision and consistency with data handling. Updated tests, models, and utilities accordingly to reflect this change.

* Improve JsonBlobParserTest by adding assertions
Replaced unnecessary print statement with proper assertions to validate parsed values. Enhanced test coverage by adding a new case for index 0 to ensure functionality correctness.

* Change double to Double in JsonBlobParser methods
Updated methods to return Double instead of primitive double to handle potential null values gracefully. This is more consistent with the previous Float implementation.

+++ services/picsure-dictionary/src/main/java/edu/harvard/dbmi/avillach/dictionary/concept/model/ContinuousConcept.java
@@ -11,7 +11,7 @@ import java.util.Objects;
 public record ContinuousConcept(
     String conceptPath, String name, String display, String dataset, String description, boolean allowFiltering,
 
-    @Nullable Float min, @Nullable Float max, String studyAcronym, Map<String, String> meta, @Nullable List<Concept> children,
+    @Nullable Double min, @Nullable Double max, String studyAcronym, Map<String, String> meta, @Nullable List<Concept> children,
 
     @Nullable Concept table,
 
@@ -19,8 +19,8 @@ public record ContinuousConcept(
 ) implements Concept {
 
     public ContinuousConcept(
-        String conceptPath, String name, String display, String dataset, String description, boolean allowFiltering, @Nullable Float min,
-        @Nullable Float max, String studyAcronym, Map<String, String> meta, @Nullable List<Concept> children
+        String conceptPath, String name, String display, String dataset, String description, boolean allowFiltering, @Nullable Double min,
+        @Nullable Double max, String studyAcronym, Map<String, String> meta, @Nullable List<Concept> children
     ) {
         this(conceptPath, name, display, dataset, description, allowFiltering, min, max, studyAcronym, meta, children, null, null);
     }
@@ -37,8 +37,8 @@ public record ContinuousConcept(
     }
 
     public ContinuousConcept(
-        String conceptPath, String name, String display, String dataset, String description, boolean allowFiltering, @Nullable Float min,
-        @Nullable Float max, String studyAcronym, Map<String, String> meta
+        String conceptPath, String name, String display, String dataset, String description, boolean allowFiltering, @Nullable Double min,
+        @Nullable Double max, String studyAcronym, Map<String, String> meta
     ) {
         this(conceptPath, name, display, dataset, description, allowFiltering, min, max, studyAcronym, meta, null);
     }
